+++ src/generate/register.rs
@@ -7,8 +7,8 @@ use cast::u64;
 use log::warn;
 use proc_macro2::{Ident, Punct, Spacing, Span, TokenStream};
 
-use crate::errors::*;
 use crate::util::{self, ToSanitizedSnakeCase, ToSanitizedUpperCase, U32Ext};
+use anyhow::{anyhow, Result};
 
 pub fn render(
     register: &Register,
@@ -26,7 +26,7 @@ pub fn render(
     let rsize = register
         .size
         .or(defs.size)
-        .ok_or_else(|| format!("Register {} has no `size` field", register.name))?;
+        .ok_or_else(|| anyhow!("Register {} has no `size` field", register.name))?;
     let rsize = if rsize < 8 {
         8
     } else if rsize.is_power_of_two() {
@@ -38,7 +38,7 @@ pub fn render(
     let description = util::escape_brackets(
         util::respace(&register.description.clone().unwrap_or_else(|| {
             warn!("Missing description for register {}", register.name);
-            "".to_string()
+            Default::default()
         }))
         .as_ref(),
     );
@@ -264,7 +264,7 @@ pub fn fields(
                             .map(|element| element.parse::<u32>())
                             .eq((first..de.dim + first).map(Ok));
                         if !sequential_indexes {
-                            return Err(format!("unsupported array indexes in {}", f.name))?;
+                            return Err(anyhow!("unsupported array indexes in {}", f.name))?;
                         }
                         (first, None)
                     } else {
@@ -282,7 +282,7 @@ pub fn fields(
             }
             Field::Single(_) => {
                 if f.name.contains("%s") {
-                    return Err(format!("incorrect field {}", f.name))?;
+                    return Err(anyhow!("incorrect field {}", f.name))?;
                 }
                 None
             }
@@ -678,7 +678,7 @@ impl Variant {
             .filter(|field| field.name.to_lowercase() != "reserved")
             .map(|ev| {
                 let value = u64(ev.value.ok_or_else(|| {
-                    format!("EnumeratedValue {} has no `<value>` field", ev.name)
+                    anyhow!("EnumeratedValue {} has no `<value>` field", ev.name)
                 })?);
 
                 Ok(Variant {
@@ -896,9 +896,10 @@ fn lookup_in_fields<'f>(
     if let Some(base_field) = fields.iter().find(|f| f.name == base_field) {
         lookup_in_field(base_evs, None, None, base_field)
     } else {
-        Err(format!(
+        Err(anyhow!(
             "Field {} not found in register {}",
-            base_field, register.name
+            base_field,
+            register.name
         )
         .into())
     }
@@ -923,12 +924,17 @@ fn lookup_in_peripheral<'p>(
         {
             lookup_in_field(base_evs, Some(base_register), base_peripheral, field)
         } else {
-            Err(format!("No field {} in register {}", base_field, register.name).into())
+            Err(anyhow!(
+                "No field {} in register {}",
+                base_field,
+                register.name
+            ))?
         }
     } else {
-        Err(format!(
+        Err(anyhow!(
             "No register {} in peripheral {}",
-            base_register, peripheral.name
+            base_register,
+            peripheral.name
         )
         .into())
     }
@@ -953,7 +959,11 @@ fn lookup_in_field<'f>(
         }
     }
 
-    Err(format!("No EnumeratedValues {} in field {}", base_evs, field.name).into())
+    Err(anyhow!(
+        "No EnumeratedValues {} in field {}",
+        base_evs,
+        field.name
+    ))?
 }
 
 fn lookup_in_register<'r>(
@@ -973,9 +983,10 @@ fn lookup_in_register<'r>(
     }
 
     match matches.first() {
-        None => Err(format!(
+        None => Err(anyhow!(
             "EnumeratedValues {} not found in register {}",
-            base_evs, register.name
+            base_evs,
+            register.name
         )
         .into()),
         Some(&(evs, field)) => {
@@ -990,10 +1001,11 @@ fn lookup_in_register<'r>(
                 ))
             } else {
                 let fields = matches.iter().map(|(f, _)| &f.name).collect::<Vec<_>>();
-                Err(format!(
+                Err(anyhow!(
                     "Fields {:?} have an \
                      enumeratedValues named {}",
-                    fields, base_evs
+                    fields,
+                    base_evs
                 )
                 .into())
             }
@@ -1019,7 +1031,7 @@ fn lookup_in_peripherals<'p>(
             peripheral,
         )
     } else {
-        Err(format!("No peripheral {}", base_peripheral).into())
+        Err(anyhow!("No peripheral {}", base_peripheral))?
     }
 }
 
@@ -1037,21 +1049,19 @@ fn periph_all_registers<'a>(p: &'a Peripheral) -> Vec<&'a Register> {
     }
 
     loop {
-        let b = rem.pop();
-        if b.is_none() {
-            break;
-        }
-
-        let b = b.unwrap();
-        match b {
-            RegisterCluster::Register(reg) => {
-                par.push(reg);
-            }
-            RegisterCluster::Cluster(cluster) => {
-                for c in cluster.children.iter() {
-                    rem.push(c);
+        if let Some(b) = rem.pop() {
+            match b {
+                RegisterCluster::Register(reg) => {
+                    par.push(reg);
+                }
+                RegisterCluster::Cluster(cluster) => {
+                    for c in cluster.children.iter() {
+                        rem.push(c);
+                    }
                 }
             }
+        } else {
+            break;
         }
     }
     par